Add command line options to test_nfc.
By default with no command line options test_nfc will initialise and
deinitialise the NFC stack.

The -n option causes test_nfc to continuously scan for targets and
read NDEF messages from them.

The -r option causes the NFC stack to be deinitialised and
reinitialised in the event of a failure to connect to the remote
device.