dynamically add hooks from libc
The hook table contains functions that need
parameter conversion(like FILE*).

If a symbol is not found in the hook table, it's searched
in libc.so.6. If it's not found there, then we fallback to bionic.

The pthread functions that we don't have in the hook table,
when called by client code, will cause a crash.

For debugging purposes, one can use HYBRIS_FROM_BIONIC to
force a symbol to be used from Bionic.

So far, I only need longjmp and setjmp from Bionic, everything else
being called from (e)glibc.

Tested on nexus 4 as:
  HYBRIS_FROM_BIONIC="setjmp:longjmp" \
  HYBRIS_EGLPLATFORM=hwcomposer \
  ./ARMHF/bin/test_hwcomposer

Signed-off-by: Adrian Negreanu <adrian.m.negreanu@intel.com>