egl: wayland: Abort the application if the wayland display becomes unusable
If there's an error while dispatching the wayland queue, the display might
be fatally busted. We can't recover from that so it's better to abort directly
rather than get stuck in strange situations or crash a while later.

Should make debugging the problem a bit easier when the initial error is
clear and not eg. deadlocking after we don't get frame events anymore.

Signed-off-by: Kalle Vahlman <kalle.vahlman@movial.com>