egl: Fix calling down to eglCreateImageKHR() with proper arguments
We do want to call eglCreateImageKHR() with EGL_NO_CONTEXT, but ONLY
when the render target is EGL_NATIVE_BUFFER_ANDROID. Otherwise we'll
either not get a valid image even if we should or worse, end up crashing
if the drivers don't handle it properly. Credit goes to Menghan Cheng for
tracing this problem to its source.

While we are touching the code, change it so the ws has an opportunity to
change all the parameters should it have specific needs for things like this.

Signed-off-by: Kalle Vahlman <kalle.vahlman@movial.com>