Implement the EGL_WL_create_wayland_buffer_from_image extension
This extension allows a nested compositor to create a client side
buffer for a server side buffer resource, and to post it to the parent
compositor on a subsurface, outsourcing compositing of the client surface
to the parent.
We also wrap EGLImageKHR with an internal struct to store the buffer that
image was created from, which we use later to create the wl_buffer.