[egl] eglGetProcAddress should return symbols of chosen EGL config
Previously we just looked up the process-global namespace for the functions,
so looking up symbols only worked if the process also had the right
libGLES*.so loaded.

EGL on other platforms doesn't need this, so make sure that we load the right
library dynamically and return symbols from it. The library loaded is still the
libhybris (glibc-based) one due to the hardfp calling convention.

This also avoids issues where eglGetProcAddress() would return any symbols in
the global symbol namespace of the dynamic linker.