Get rid of workaround allowing gecko to pass invalid pointers to EGL.
This basically reverts 4c6d9966187e561a2e6b821aee0e90c11f413e9a.
Acording to Gunnar Sletta the patch was a workaround for a bug in gecko.
Since the problem was fixed there is no point in keeping it around.
Passing invalid pointers to EGL functions should not be an acceptable
behaviour.