Propagate eglQueryString error back to the caller instead of crashing.
The EGL spec defines the conditions in which the function can fail. Some
applications like chromium can actually cope with the fact that an early
eglQueryString function call failed and can retry it at a later time.
Unfortunately current implementation of EGL in libhybris simply aborts
exeution of the app in such case. This patch removes such bogus asserts
propagating the NULL return value to the caller instead.