Fix resolving the GLEsv2 functions at link time
The issue was addressed in 295d159 but that commit had to be reverted by
c1099bd because of a crash when trying to dlopen the wrapped libGLESv2.so.
Instead, if a gl* function is resolved when we don't have yet a handle for
the library, use a wrapper calling the internal function.