wifi: fix hardware_legacy/wifi.h not found
Android headers were supposed to be installed to /usr/include. When it's
not the case, even with a correct C include path set for Android
headers, the build still fails with "hardware_legacy/wifi.h not found".

This patch removes the redundant level of directory from wifi related
include directives.