hardware: provide our own hardware.h
To be able to stay independent of a specific Android header set we need
to have a independent libhardware too. The libhardware API is stable and
doesn't really change as it's just a versioning mechanism around dlopen
and module initialization. We now ship our own hardware.h providing the
API for the library and fail at configuration time if we find another
hardware.h in the build system.