compat: camera: Fix a use-after-free when creating a new CameraControl after deleting an old one
CameraControl is reference counted, and android_camera_connect_to stores
a strong reference to it in a global pointer. However, android_camera_delete
just deletes the CameraControl instance, which results in a UAF when
calling android_camera_connect_to again.

This patch drops the global reference to CameraControl, and relies on the
fact that CameraControl is kept alive because android::Camera holds a
strong reference to it as its listener. As CameraControl also owns a strong
reference to android::Camera, this creates a reference loop that we break
in android_camera_delete