debian: adjust our build depends depending which distro we build for
On vivid we still depend on gcc 4.7 and can't switch yet to a newer
toolchain as this causes silent ABI regressions we didn't ruled out
yet. On newer distro 4.7 was removed from the archive so we have to
used a newer toolchain.

This change auto generates debian/control from a template in debian/
control.in and adds the build-dep on gcc-4.7/g++-4.7 on demand.