compat: media: don't put Android 5.x as a requirement for MediaCodecSource
We will most likely get the case that a device running Android 4.x
will get a backport of the MediaCodecSource class. In that case we
should still support it through the media compat API. The device
running Android 4.x then only has to specify a

 BOARD_HAS_MEDIA_CODEC_SOURCE := true

in its BoardConfig.mk and we will add support for it.