hybris: common: enable linker overrides only by configure option
This way we can selectively enable/disable them per build architecture
in the distribution packaging. Intent behind this is to let devices we
may migrate in the future from armhf to arm64 stay on the linker
version they're forced to to not change anything on what we have in
production but fix any bugs with a newer linker version only for arm64.