hooks: fix arm64 shared pthread ptr implementation
libhybris previously stores memory addresses of shared mutexes in a
unsigned integer. When running on arm64 or any other __LP64__ systems,
this results in wrong pointer addresses returned. In this patch we
re-type to storage type to uintptr_t, which guarantees sufficient space
for holding a memory address, and then corretly set HYBRIS_SHM_MASK and
HYBRIS_SHM_MASK_TOP to valid values on both systems.