hybris: use as few hooks as possible and initialize libc.
can be enabled by adding --enable-initialize-bionic to the automake based build

unfortunately testing has shown that pthread_gettid_np does not function properly
but for this method it is mandatory that this function works.
this stores a map of active PIDs and uses them to simulate pthread_gettid_np.

WARNING: when this is enabled a patch is required to android_bionic! reffering to https://github.com/krnlyng/android_bionic/commit/aaf4ebae112028f6b02cb2c64123e914d0422111