utils: hardware_legacy/wifi.h needs to be extracted for AOSP 4.1+
This header was previously extracted with Android 6 only but it is available since 4.1.
The wifi module compilation is not set to be compiled only for Android 6.

That permit to fix compilation issue with older version than 6 on wifi.c :
wifi.c:23:34: fatal error: hardware_legacy/wifi.h: No such file or directory