Make compat libraries build for android-7
The ICamera interface was moved to android::hardware::ICamera. The
trunk of the changes is related to forward declarations in the hybris
code that leads to confusing error messages. Makefile changes are
related to the three evolving C++ standard libraries in the build tree.
Maybe we can get rid of stlport - and banish bionics broken headers..

Besides those mechanical changes there are also two changes in the
media recording and camaera callbacks that might require changes in the
future. A method called requestIDRFrame was removed from
MediaSourcePrivate and is only available in a private member. This
member gets triggered via onStart and onResume. So maybe it is not
necessary. Then there is another callback in CameraListener, which gets
a native_handle_t - maybe that is the current frame?

Change-Id: I56135f14131fc57ae56e45355b2fa7009be209e9