Ensure that wayland window operations don't crash or hang on NULL m_window.
The free_callback() routine will be called on the main thread, putting
m_window to NULL. Other operations may still be pending on the rendering
thread and will segfault when accessing m_window. Prevent this by locking
access to m_window and ensuring that m_window being NULL ends by a quick
return. For instance eglSwapBuffers() will first call prepareSwap() which
will be locked during the call, then lock is released and the underlying
Android _eglSwapBuffers() is performed, and finally finishSwap() is
called and locking. So it's not only necessary to lock the nullification
of m_window, but also to ensure that finishSwap() can be called with
m_window being NULL, since the nullification may happen during the time
the process is unlocked in underlying Android call.