jb/mm/n: HYBRIS_ENABLE_LINKER_DEBUG_MAP enables Android libs in gdb
There are still some conflicts between android libs and glibc libs in the
gdb lib map.

Very likely, in most cases, the developer only wants to debug the glibc
related code, and is not interested in seeing the Android stack.
Therefore, introduce an env. variable and disable linking of android
libs by default.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>