hybris: hook __cxa_thread_atexit
This function is used by the compiler to destruct `thread_local`
variables at thread's exit. Virtually no one uses it, but I planed to
use to solve libmedia_compat_layer's problem when using remote codec.

This implementation wraps dtor so that we can do reference counting.
This prevents memory leak when a library is dlclose'd, at the expense
of more memory usage and more code to execute.