hooks: hook __cxa_atexit ourself to prevent segfault at exit
If an Android library which place a __cxa_atexit() hook is not
dlclose()'ed before libhybris-common is unloaded, the hook will be
called when the program exit. This can cause segfault if the hook
refers to libhybris'es hooked symbol.

Record libraries that placed hooks, and then call them (through
__cxa_finallize) when we're exiting.

As hooks.c is growing, and because this is easier to implement in C++,
I've decided create a subdirectory specifically to contains hooks in the
future. Files in this subdirectory will be compiled with -fvisibility=
hidden to reduce global symbol pollution.