egl: provide functions for EGL_EXT_platform_base
libglvnd will advertise EGL_EXT_platform_base when one of the vendor
libraries support the extension. And most of the time, we're installed
together with Mesa, which support the extension, so we're forced to
support it or some client (e.g. Weston nested) will not work properly.

See https://gitlab.freedesktop.org/glvnd/libglvnd/-/issues/214