[egl] Set buffer dimensions on window resize. JB#52698
Some EGL implementations rightfully doesn't ask the default width
and height of the NativeWindow until transform hint changes, so
SET_BUFFERS_DIMENSIONS is never called. This results in the EGL
surface not being resized when wl_egl_window_resize() is called on
the window.

With this commit WaylandNativeWindow::resize() will set both default
and current dimensions and each buffer will be reallocated as soon as
it is dequeued.