hooks: Reset reference count of condition variable before pthread_cond_destroy.
Bionic and glibc implementations of pthread_cond_destroy are different.
Bionic implementation is trivial and does not block whereas the glibc
implementation requires that there is no thread waiting for the condition
variable when it is destroyed and bionic code does not always follow this
requirement which can cause a deadlock because two or more threads
are waiting for the same condition variable.