hybris: egl: Provide eglCreatePlatformWindowSurface. Since https://github.com/libhybris/libhybris/commit/a1bf776f4c06caba8078c7335deb766b3a77fdf1 the eglGetPlatformDisplay() function has been implemented. This function is part of EGL 1.5.
Some platforms rely on the existence of this function to detect if EGL 1.5 is supported [^1].
However, libhybris does not implement all EGL 1.5 functions, this provides the implementation for the eglCreatePlatformWindowSurface() function to fix a compilation issue with Qt [^2].

It's worth noting that the eglCreatePlatformWindowSurface() function and the eglCreateWindowSurface() function are not identical.
Similar to https://github.com/libhybris/libhybris/commit/362efd97cd5b18f580c402d3fc14555f1ef6e31e we might need to implement additional type conversions at a later stage.

[^1]: https://github.com/qt/qtwayland/blob/5.15.2/src/client/configure.json#L183
[^2]: https://github.com/qt/qtwayland/blob/5.15.2/src/hardwareintegration/client/wayland-egl/qwaylandglcontext.cpp#L356
Signed-off-by: Darrel Griët <dgriet@gmail.com>